Criando API do Relatorio e retornando informação do Relatorio na tela

+++ back/src/main/java/com/avaliacao/ouvidoria/domain/model/Relatorio.java
@@ -42,7 +42,7 @@ public class Relatorio {
 	
 	@NotBlank
 	@Column
-	private Integer mediaTotal;
+	private double mediaTotal;
 	
 	@NotBlank
 	@Column
